+++ chartlets.js/src/lib/components/Plot.tsx
@@ -1,40 +1,30 @@
-import { VegaLite, type VisualizationSpec } from "react-vega";
+import { VegaLite } from "react-vega";
 
 import { type ComponentState } from "@/lib/types/state/component";
 import type { ComponentProps } from "@/lib/component/Component";
+import { useSignalListeners } from "@/lib/hooks";
+import type { TopLevelSpec } from "vega-lite";
 
 interface PlotState extends ComponentState {
   chart?:
-    | (VisualizationSpec & {
-        datasets?: Record<string, unknown>; // Add the datasets property
-      })
-    | null
-    | undefined;
+    | TopLevelSpec // This is the vega-lite specification type
+    | null;
 }
 
 interface PlotProps extends ComponentProps, PlotState {}
 
 export function Plot({ type, id, style, chart, onChange }: PlotProps) {
+  const signalListeners = useSignalListeners(chart, type, id, onChange);
+
   if (!chart) {
     return <div id={id} style={style} />;
   }
-  const { datasets, ...spec } = chart;
-  const handleSignal = (_signalName: string, value: unknown) => {
-    if (id) {
-      return onChange({
-        componentType: type,
-        id: id,
-        property: "points",
-        value: value,
-      });
-    }
-  };
+
   return (
     <VegaLite
-      spec={spec}
-      data={datasets}
+      spec={chart}
       style={style}
-      signalListeners={{ onClick: handleSignal }}
+      signalListeners={signalListeners}
       actions={false}
     />
   );