+++ chartlets.js/src/lib/types/state/vega.ts
@@ -1,0 +1,1 @@
+export type SignalHandler = (signalName: string, value: unknown) => void;